+++ cmd/main.go
@@ -17,6 +17,9 @@ func main() {
 	// Perform config validation
 	config.Validate()
 
+	// initialize the code store
+	internal.InitCodeStore()
+
 	// Build server
 	server := internal.NewServer()
