+++ internal/server.go
@@ -1,7 +1,6 @@
 package tfa
 
 import (
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/url"
@@ -187,16 +186,19 @@ func (s *Server) AuthCallbackHandler() http.HandlerFunc {
 		logger.Debug("User FirstName--------------------------------------------------->" + user.FirstName)
 		logger.Debug("User LastName--------------------------------------------------->" + user.LastName)
 		// Generate cookie
-		http.SetCookie(w, MakeCookie(r, user.Email))
-		http.SetCookie(w, MakeCookie(r, user.ID))
+		// http.SetCookie(w, MakeCookie(r, user.Email))
+		cookie := MakeCookie(r, user.ID)
+		http.SetCookie(w, cookie)
 
-		cookie, err := MakeUserCookie(r, fmt.Sprintf("%s|%s|%s", user.Email, user.FirstName, user.LastName))
+		saveCode(cookie.Value, r.URL.Query().Get("code"))
+
+		userInfoCookie, err := MakeUserCookie(r, fmt.Sprintf("%s|%s|%s", user.Email, user.FirstName, user.LastName))
 		if err != nil {
 			logger.Errorf("MakeUserCookie: %v", err)
 			http.Error(w, "Not authorized", http.StatusInternalServerError)
 			return
 		}
-		http.SetCookie(w, cookie)
+		http.SetCookie(w, userInfoCookie)
 
 		logger.WithFields(logrus.Fields{
 			"user_Email": user.Email,
@@ -205,64 +207,6 @@ func (s *Server) AuthCallbackHandler() http.HandlerFunc {
 
 		// Redirect
 		http.Redirect(w, r, redirect, http.StatusTemporaryRedirect)
-	}
-}
-
-// GetUserInfoHandler Handles auth callback request
-func (s *Server) GetUserInfoHandler() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		// Logging setup
-		logger := s.logger(r, "GetUserInfoHandler", "default", "Handling callback")
-
-		providerName := "oidc"
-
-		// Get provider
-		p, err := config.GetConfiguredProvider(providerName)
-		if err != nil {
-			logger.WithFields(logrus.Fields{
-				"error": err,
-				// "csrf_cookie": c,
-				"provider": providerName,
-			}).Warn("Invalid provider in csrf cookie")
-			http.Error(w, "Not authorized", http.StatusUnauthorized)
-			return
-		}
-
-		// if code is missing, redirect user to auth handler with callbackurl as
-		if r.URL.Query().Get("code") == "" {
-			redirectPath := config.Path + "/userinfo"
-			s.authRedirect(logger, w, r, p, redirectPath)
-			return
-		}
-
-		redirectURI := &url.URL{
-			Scheme: r.Header.Get("X-Forwarded-Proto"),
-			Host:   r.Host,
-			Path:   config.Path,
-		}
-
-		user, err := p.GetUserFromCode(r.URL.Query().Get("code"), redirectURI.String())
-		if err != nil {
-			logger.Errorf("GetUserFromCode: %v", err)
-			http.Error(w, "Not authorized", http.StatusUnauthorized)
-			return
-		}
-
-		logger.Debug("User ID--------------------------------------------------->" + user.ID)
-		logger.Debug("User Email--------------------------------------------------->" + user.Email)
-		logger.Debug("User FirstName--------------------------------------------------->" + user.FirstName)
-		logger.Debug("User LastName--------------------------------------------------->" + user.LastName)
-
-		if err := json.NewEncoder(w).Encode(user); err != nil {
-			logger.Errorf("GetUserFromCode: %v", err)
-			http.Error(w, "failed to encode user information", http.StatusInternalServerError)
-			return
-		}
-
-		logger.WithFields(logrus.Fields{
-			"user_Email": user.Email,
-			"user_ID":    user.ID,
-		}).Infof("Generated auth cookie")
 	}
 }
